move offending class ArchivedProcessFailure to the public API and adjust all references, update exported API version
